eval/requirements.txt
mteb
python-dotenv
streamlit
setproctitle
mteb[gritlm]
mteb[jina]